+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+### Changed
+- disabled => isDisabled, autofocus => isAutofocus are now boolean props
+- _open
+- _disabled
+- _has
+
 ## [1.1.2] - 2021-03-06
 ### Fixed
 - Deps security issues